+++ IncrementalInference/src/services/MaxMixture.jl
@@ -11,7 +11,7 @@ struct MaxMixture <: AbstractMaxMixtureSolver
   choice::Base.RefValue{Int}
 end
 
-function getMeasurementParametric(s::Mixture{N, F, S, T}) where {N, F, S, T}
+function getMeasurementParametric(s::Mixture)
   meas = map(c -> getMeasurementParametric(c)[1], values(s.components))
   iΣ = map(c -> getMeasurementParametric(c)[2], values(s.components))
   return meas, iΣ